+++ src/HotelPosSystem/MainForm.cs
@@ -1,101 +1,105 @@
+using System.Globalization;
+
 using HotelPosSystem.Entities;
+
+using Microsoft.EntityFrameworkCore;
 
 namespace HotelPosSystem {
     public partial class MainForm : Form {
-        private readonly uint[] _occupiedRooms;
-        private readonly Label[] _roomTypeCounterTexts;
-
         public MainForm() {
             InitializeComponent();
 
             using HotelDbContext databaseContext = new();
             DatabaseUtilities.SetUpDatabase(databaseContext);
 
-            _occupiedRooms = new uint[databaseContext.RoomTypes.Count()];
-            _roomTypeCounterTexts = new Label[_occupiedRooms.Length];
-
-            FlowLayoutPanel verticalLayoutPanel = new() {
+            FlowLayoutPanel bookingList = new() {
+                Name = "bookingList",
                 FlowDirection = FlowDirection.TopDown,
-                AutoSize = true
+                AutoScroll = true,
+                WrapContents = false,
+                Dock = DockStyle.Fill
             };
 
-            Label occupiedRoomsHeading = new() {
-                Text = "Occupied Rooms",
-                Font = new Font(Font.FontFamily, 14),
-                AutoSize = true
-            };
-            verticalLayoutPanel.Controls.Add(occupiedRoomsHeading);
-            RoomType[] roomTypes = databaseContext.RoomTypes.ToArray();
-            for (int i = 0; i < roomTypes.Length; i++) {
-                FlowLayoutPanel row = CreateCounterRow(i);
-                verticalLayoutPanel.Controls.Add(row);
-            }
-
-            Controls.Add(verticalLayoutPanel);
-        }
-
-        private void OnIncrementButtonClicked(int index) {
-            _occupiedRooms[index]++;
-            UpdateRoomTypeCounterText(index);
-        }
-
-        private void OnDecrementButtonClicked(int index) {
-            if (_occupiedRooms[index] > 0) {
-                _occupiedRooms[index]--;
-                UpdateRoomTypeCounterText(index);
-            }
-        }
-
-        private void OnResetButtonClicked(int index) {
-            _occupiedRooms[index] = 0;
-            UpdateRoomTypeCounterText(index);
-        }
-
-        private void UpdateRoomTypeCounterText(int index) {
-            RoomType roomType = GetRoomType(index);
-            string newText = $"{roomType.Name}: {_occupiedRooms[index]}";
-            _roomTypeCounterTexts[index].Text = newText;
-        }
-
-        private FlowLayoutPanel CreateCounterRow(int index) {
-            RoomType roomType = GetRoomType(index);
-            FlowLayoutPanel result = new() {
-                FlowDirection = FlowDirection.LeftToRight,
+            Label bookingListHeading = new() {
+                Text = "Existing bookings",
+                Font = new Font(Font.FontFamily, 18),
                 AutoSize = true,
-                Name = $"{roomType.Name} Row"
+                UseCompatibleTextRendering = true
             };
+            bookingList.Controls.Add(bookingListHeading);
 
-            Label counterText = new() {
-                Name = "roomTypeCounterText",
-                AutoSize = true
-            };
-            _roomTypeCounterTexts[index] = counterText;
-            UpdateRoomTypeCounterText(index);
-            result.Controls.Add(counterText);
 
-            Button decrementButton =
-              CreateButton("decrementButton", "-",
-              (object? sender, EventArgs eventArgs) => OnDecrementButtonClicked(index));
-            result.Controls.Add(decrementButton);
+            Booking[] bookings = databaseContext.Bookings
+                .Include(booking => booking.Customer)
+                .Include(booking => booking.Room)
+                .OrderBy(booking => booking.StartDate)
+                .ToArray();
 
-            Button incrementButton =
-                CreateButton("incrementButton", "+",
-                (object? sender, EventArgs eventArgs) => OnIncrementButtonClicked(index));
-            result.Controls.Add(incrementButton);
-
-            Button resetButton =
-                CreateButton("resetButton", "Reset",
-                (object? sender, EventArgs eventArgs) => OnResetButtonClicked(index));
-            result.Controls.Add(resetButton);
-
-            return result;
+            foreach (Booking booking in bookings) {
+                bookingList.Controls.Add(CreateListItem(booking));
+            }
+            Controls.Add(bookingList);
         }
 
-        private static RoomType GetRoomType(int index) {
-            using HotelDbContext databaseContext = new();
-            return databaseContext.RoomTypes
-                .OrderBy(roomType => roomType.Id)
-                .ToArray()[index];
+        private static FlowLayoutPanel CreateListItem(Booking booking) {
+            FlowLayoutPanel flowLayoutPanel = new() {
+                FlowDirection = FlowDirection.TopDown,
+                AutoSize = true,
+                Margin = new Padding(0, top: 30, 0, 0)
+            };
+
+            AddLabel(flowLayoutPanel, "customerName" + booking.Id, "Name: " + booking.Customer.FullName);
+            AddLabel(flowLayoutPanel, "emailAddress" + booking.Id, "Email address: " + booking.Customer.EmailAdress);
+            AddLabel(flowLayoutPanel, "phoneNumber" + booking.Id, "Phone number: " + booking.Customer.PhoneNumber);
+            AddLabel(flowLayoutPanel, "roomName" + booking.Id, "Room: " + booking.Room.Name);
+
+            // "o" date format corresponds to the ISO 8601 standard
+            // https://learn.microsoft.com/en-us/dotnet/standard/base-types/standard-date-and-time-format-strings
+            string startDate = booking.StartDate.ToString("o", CultureInfo.InvariantCulture);
+            string endDate = booking.EndDate.ToString("o", CultureInfo.InvariantCulture);
+            AddLabel(flowLayoutPanel, "dates" + booking.Id,
+                               $"Dates: {startDate} to {endDate}");
+
+            AddCheckBoxAndLabel(flowLayoutPanel, "paidFor" + booking.Id, booking.IsPaidFor, "", "Has paid:");
+            AddCheckBoxAndLabel(flowLayoutPanel, "checkedIn" + booking.Id, booking.IsCheckedIn, "", "Has checked in:");
+
+            if (booking.Comment is not null) {
+                AddLabel(flowLayoutPanel, "comment" + booking.Id, "Comment: " + booking.Comment);
+            }
+
+            return flowLayoutPanel;
+        }
+
+        private static void AddCheckBoxAndLabel(Panel container, string checkBoxName, bool checkBoxState, string labelName, string labelText) {
+            FlowLayoutPanel layoutPanel = new() {
+                FlowDirection = FlowDirection.LeftToRight,
+                AutoSize = true
+            };
+
+            AddLabel(layoutPanel, labelName, labelText);
+            AddCheckBox(layoutPanel, checkBoxName, checkBoxState, isEnabled: false);
+
+            container.Controls.Add(layoutPanel);
+        }
+
+        private static void AddLabel(Panel container, string name, string text) {
+            Label label = new() {
+                Name = name,
+                Text = text,
+                UseCompatibleTextRendering = true,
+                AutoSize = true
+            };
+            container.Controls.Add(label);
+        }
+
+        private static void AddCheckBox(Panel container, string name, bool isChecked, bool isEnabled) {
+            CheckBox checkBox = new() {
+                Name = name,
+                Checked = isChecked,
+                Enabled = isEnabled,
+                UseCompatibleTextRendering = true
+            };
+            container.Controls.Add(checkBox);
         }
 
         private static Button CreateButton(string name, string text, EventHandler? onClicked) {