+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### Added
 
-- Occupied room counter for twin room 
+- List of bookings
+
+### Removed
+
+- Occupied room counters for single and double rooms with reset buttons
 
 ## 0.1.0 - 2023-12-07